requirements.txt
Django==2.2.6
djangorestframework==3.9.4
